Move search input and logic to proper file
