+++ front/app/containers/ProjectsIndexPage/index.tsx
@@ -1,14 +1,10 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 
 // components
 import ContentContainer from 'components/ContentContainer';
 import ProjectAndFolderCards from 'components/ProjectAndFolderCards';
 import CityLogoSection from 'components/CityLogoSection';
 import ProjectsIndexMeta from './ProjectsIndexMeta';
-import SearchInput from 'components/UI/SearchInput';
-
-// history
-import { useSearchParams } from 'react-router-dom';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
@@ -65,33 +61,7 @@ const PageTitle = styled.h1`
   `}
 `;
 
-const StyledSearchInput = styled(SearchInput)`
-  margin-bottom: 40px;
-`;
-
 const ProjectsIndex = () => {
-  const [search, setSearch] = useState('');
-  const [searchParams, setSearchParams] = useSearchParams();
-
-  useEffect(() => {
-    const focusSearch = searchParams.get('focusSearch');
-    // it's a string from the query param
-    if (focusSearch === 'true') {
-      // this should be a ref to the input element
-      // it's null on page load
-      const searchInput = document.getElementById('search-input');
-      if (searchInput) {
-        searchInput.focus();
-      }
-      searchParams.delete('focusSearch');
-      setSearchParams(searchParams);
-    }
-  }, [searchParams, setSearchParams]);
-
-  const handleSearchOnChange = (search: string) => {
-    setSearch(search);
-  };
-
   return (
     <>
       <ProjectsIndexMeta />
@@ -100,15 +70,11 @@ const ProjectsIndex = () => {
           <PageTitle>
             <FormattedMessage {...messages.pageTitle} />
           </PageTitle>
-          <StyledSearchInput
-            onChange={handleSearchOnChange}
-            a11y_numberOfSearchResults={1}
-          />
+
           <ProjectAndFolderCards
             showTitle={false}
             layout="threecolumns"
             publicationStatusFilter={['published', 'archived']}
-            search={search}
           />
         </StyledContentContainer>
         <CityLogoSection />
